admt: Updated CNV change method - Changed CNV page to uint8_t - Added registry read for rotation, angle, count, temp channels - Added disable state for lineedit - Changed error value for double to UINT32_MAX

+++ plugins/admt/include/admt/widgets/registerblockwidget.h
@@ -47,7 +47,7 @@ public:
 
 	QPushButton *m_readButton, *m_writeButton;
 
-	RegisterBlockWidget(QString header, QString description, uint32_t address, uint32_t cnvPage,
+	RegisterBlockWidget(QString header, QString description, uint32_t address, uint8_t cnvPage,
 			    RegisterBlockWidget::ACCESS_PERMISSION accessPermission, QWidget *parent = nullptr);
 	virtual ~RegisterBlockWidget();
 	QPushButton *readButton();
